Refactor keyboard builders
