+++ bot/keyboards/builders/__init__.py
@@ -1,297 +1,9 @@
-from telegram import ReplyKeyboardMarkup
+"""Keyboard builder helper functions."""
 
-from ...utils.formatting import arabic_ordinal, format_lecturer_name
-
-TERM_MENU_SHOW_SUBJECTS = "📖 عرض المواد"
-TERM_MENU_PLAN = "🗂 الخطة الدراسية"
-TERM_MENU_LINKS = "🔗 روابط المجموعات والقنوات"
-TERM_MENU_ADV_SEARCH = "🔎 البحث المتقدم"
-
-BACK = "🔙 العودة"
-BACK_TO_LEVELS = "🔙 العودة لقائمة المستويات"
-BACK_TO_SUBJECTS = "🔙 العودة لقائمة المواد"
-
-FILTER_BY_YEAR = "📂 حسب السنة"
-FILTER_BY_LECTURER = "👨‍🏫 حسب المحاضر"
-LIST_LECTURES = "📚 عرض المحاضرات"
-YEAR_MENU_LECTURES = "📚 المحاضرات"
-
-SECTION_LABELS = {
-    "theory": "📚 نظري",
-    "discussion": "💬 مناقشة",
-    "lab": "🧪 عملي",
-    "syllabus": "📄 المفردات الدراسية",
-    "apps": "📱 تطبيقات",
-}
-
-CATEGORY_TO_LABEL = {
-    "lecture": "📄 ملف المحاضرة",
-    "slides": "📑 سلايدات المحاضرة",
-    "audio": "🎧 تسجيل صوتي",
-    "video": "🎥 تسجيل فيديو",
-    "board_images": "🖼️ صور السبورة",
-    "external_link": "🔗 روابط خارجية",
-    "exam": "📝 الامتحانات",
-    "booklet": "📘 الملازم",
-    "summary": "🧾 ملخص",
-    "notes": "🗒️ ملاحظات",
-    "simulation": "🧪 محاكاة",
-    "mind_map": "🗺️ خرائط ذهنية",
-    "transcript": "⌨️ تفريغ صوتي",
-    "related": "📎 ملفات ذات صلة",
-}
-
-LIST_LECTURES_FOR_LECTURER = "📚 محاضرات هذا المحاضر"
-
-
-def _rows(items: list[str], cols: int = 2) -> list[list[str]]:
-    """Split items into rows with a fixed number of columns, skipping empties."""
-    items = [i for i in items if i]
-    keyboard, row = [], []
-    for item in items:
-        row.append(item)
-        if len(row) == cols:
-            keyboard.append(row)
-            row = []
-    if row:
-        keyboard.append(row)
-    return keyboard
-
-
-def generate_levels_keyboard(levels: list) -> ReplyKeyboardMarkup:
-    """Display available levels."""
-    names = [name for _id, name in levels]
-    keyboard = _rows(names, cols=2)
-    keyboard.append(["🔙 العودة للقائمة الرئيسية"])
-    return ReplyKeyboardMarkup(
-        keyboard=keyboard,
-        resize_keyboard=True,
-        input_field_placeholder="إختر المستوى الدراسي  ⬇️",
-    )
-
-
-def generate_terms_keyboard(terms: list) -> ReplyKeyboardMarkup:
-    """Display terms for a specific level."""
-    names = [name for _id, name in terms]
-    keyboard = _rows(names, cols=2)
-    keyboard.append([BACK])
-    return ReplyKeyboardMarkup(
-        keyboard=keyboard,
-        resize_keyboard=True,
-        input_field_placeholder="إختر الفصل الدراسي  ⬇️",
-    )
-
-
-def generate_subjects_keyboard(subjects: list) -> ReplyKeyboardMarkup:
-    """Display subjects for the current term."""
-    names = [name for (name,) in subjects]
-    keyboard = _rows(names, cols=2)
-    keyboard.append([BACK])
-    keyboard.append([BACK_TO_LEVELS])
-    return ReplyKeyboardMarkup(
-        keyboard=keyboard,
-        resize_keyboard=True,
-        input_field_placeholder="إختر المقرر الدراسي  ⬇️",
-    )
-
-
-def generate_term_menu_keyboard_dynamic(flags: dict) -> ReplyKeyboardMarkup:
-    """Create term menu buttons based on available data."""
-    items: list[str] = []
-    if flags.get("has_subjects"):
-        items.append(TERM_MENU_SHOW_SUBJECTS)
-    if flags.get("has_syllabus"):
-        items.append(TERM_MENU_PLAN)
-    if flags.get("has_links"):
-        items.append(TERM_MENU_LINKS)
-    if flags.get("has_subjects") or flags.get("has_syllabus") or flags.get("has_links"):
-        items.append(TERM_MENU_ADV_SEARCH)
-
-    keyboard = _rows(items, cols=2)
-    keyboard.append([BACK, BACK_TO_LEVELS])
-    return ReplyKeyboardMarkup(keyboard=keyboard, resize_keyboard=True)
-
-
-def build_main_menu(is_admin: bool = False) -> ReplyKeyboardMarkup:
-    keyboard = [
-        ["📚 المستويات", "🗂 الخطة الدراسية"],
-        ["🔧 البرامج الهندسية", " بحث"],
-        ["📡 القنوات والمجموعات", "🆘 مساعدة"],
-        ["📨 تواصل معنا"],
-    ]
-    if is_admin:
-        keyboard.append(["👤 إدارة المشرفين"])
-    return ReplyKeyboardMarkup(
-        keyboard=keyboard,
-        resize_keyboard=True,
-        one_time_keyboard=True,
-        input_field_placeholder="اختر خيارًا من القائمة  ⬇️",
-    )
-
-
-def generate_subject_sections_keyboard_dynamic(sections: list[str]) -> ReplyKeyboardMarkup:
-    """Display available sections for a subject."""
-    labels = [SECTION_LABELS[s] for s in sections if s in SECTION_LABELS]
-    keyboard = _rows(labels, cols=2)
-    keyboard.append([BACK])
-    return ReplyKeyboardMarkup(keyboard=keyboard, resize_keyboard=True)
-
-
-def generate_section_filters_keyboard_dynamic(
-    years_exist: bool, lecturers_exist: bool, lectures_exist: bool
-) -> ReplyKeyboardMarkup:
-    """Provide section filters for year, lecturer, and lectures."""
-    first_row: list[str] = []
-    if years_exist:
-        first_row.append(FILTER_BY_YEAR)
-    if lecturers_exist:
-        first_row.append(FILTER_BY_LECTURER)
-
-    keyboard: list[list[str]] = []
-    if first_row:
-        keyboard.append(first_row)
-    if lectures_exist:
-        keyboard.append([LIST_LECTURES])
-
-    keyboard.append([BACK, BACK_TO_SUBJECTS])
-    keyboard.append([BACK_TO_LEVELS])
-    return ReplyKeyboardMarkup(keyboard=keyboard, resize_keyboard=True)
-
-
-def generate_years_keyboard(years: list[tuple[int, str]]) -> ReplyKeyboardMarkup:
-    """Display available years for a subject/section."""
-    names = [name for _id, name in years]
-    keyboard = _rows(names, cols=2)
-    keyboard.append([BACK, BACK_TO_SUBJECTS])
-    keyboard.append([BACK_TO_LEVELS])
-    return ReplyKeyboardMarkup(keyboard=keyboard, resize_keyboard=True)
-
-
-def generate_lecturers_keyboard(lecturers: list[tuple[int, str]]) -> ReplyKeyboardMarkup:
-    """Display lecturers for the subject/section."""
-    names = [format_lecturer_name(name) for _id, name in lecturers]
-    keyboard = _rows(names, cols=2)
-    keyboard.append([BACK, BACK_TO_SUBJECTS])
-    keyboard.append([BACK_TO_LEVELS])
-    return ReplyKeyboardMarkup(keyboard=keyboard, resize_keyboard=True)
-
-
-def generate_lecture_titles_keyboard(titles: list[str]) -> ReplyKeyboardMarkup:
-    """Display lecture titles."""
-    keyboard = _rows(titles, cols=2)
-    keyboard.append([BACK, BACK_TO_SUBJECTS])
-    keyboard.append([BACK_TO_LEVELS])
-    return ReplyKeyboardMarkup(keyboard=keyboard, resize_keyboard=True)
-
-
-def generate_lecturer_filter_keyboard(
-    _years_exist: bool, lectures_exist: bool
-) -> ReplyKeyboardMarkup:
-    """Provide filters when viewing a lecturer."""
-    row: list[str] = []
-    if lectures_exist:
-        row.append(LIST_LECTURES_FOR_LECTURER)
-
-    keyboard: list[list[str]] = []
-    if row:
-        keyboard.append(row)
-    keyboard.append([BACK])
-    return ReplyKeyboardMarkup(keyboard=keyboard, resize_keyboard=True)
-
-
-def generate_year_category_menu_keyboard(
-    categories: list[str], lectures_exist: bool
-) -> ReplyKeyboardMarkup:
-    """Display categories for a given year."""
-    labels = [CATEGORY_TO_LABEL.get(c, c) for c in categories]
-    first_row: list[str] = []
-    if lectures_exist:
-        first_row.append(YEAR_MENU_LECTURES)
-
-    keyboard: list[list[str]] = []
-    if first_row:
-        keyboard.append(first_row)
-    keyboard += _rows(labels, cols=2)
-
-    keyboard.append([BACK, BACK_TO_SUBJECTS])
-    keyboard.append([BACK_TO_LEVELS])
-    return ReplyKeyboardMarkup(keyboard=keyboard, resize_keyboard=True)
-
-
-def generate_lecture_category_menu_keyboard(categories: list[str]) -> ReplyKeyboardMarkup:
-    """Display categories for a specific lecture."""
-    labels = [CATEGORY_TO_LABEL.get(c, c) for c in categories]
-    keyboard = _rows(labels, cols=2)
-    keyboard.append([BACK, BACK_TO_SUBJECTS])
-    keyboard.append([BACK_TO_LEVELS])
-    return ReplyKeyboardMarkup(keyboard=keyboard, resize_keyboard=True)
-
-
-def build_years_menu(years: list[int]) -> ReplyKeyboardMarkup:
-    """Build a simple menu for available years."""
-    labels = [str(y) for y in years]
-    keyboard = _rows(labels, cols=2)
-    keyboard.append([BACK])
-    return ReplyKeyboardMarkup(keyboard=keyboard, resize_keyboard=True)
-
-
-def build_lectures_menu(lectures: list[dict]) -> ReplyKeyboardMarkup:
-    """Build lecture list with Arabic ordinals."""
-    labels: list[str] = []
-    for item in lectures:
-        no = item.get("lecture_no")
-        title = item.get("title", "")
-        label = f"المحاضرة {arabic_ordinal(no)}"
-        if title:
-            label += f": {title}"
-        labels.append(label)
-    keyboard = _rows(labels, cols=1)
-    keyboard.append([BACK])
-    return ReplyKeyboardMarkup(keyboard=keyboard, resize_keyboard=True)
-
-
-def build_types_menu(types: list[str]) -> ReplyKeyboardMarkup:
-    """Build menu for available content types."""
-    labels = [CATEGORY_TO_LABEL.get(t, t) for t in types]
-    keyboard = _rows(labels, cols=2)
-    keyboard.append([BACK])
-    return ReplyKeyboardMarkup(keyboard=keyboard, resize_keyboard=True)
-
-
-def build_exam_menu(mid_exists: bool, final_exists: bool) -> ReplyKeyboardMarkup:
-    """Build submenu for exam models."""
-    row: list[str] = []
-    if mid_exists:
-        row.append("النصفي")
-    if final_exists:
-        row.append("النهائي")
-    keyboard: list[list[str]] = []
-    if row:
-        keyboard.append(row)
-    keyboard.append([BACK])
-    return ReplyKeyboardMarkup(keyboard=keyboard, resize_keyboard=True)
-
+from .main_menu import build_main_menu
+from .paginated import build_children_keyboard
 
 __all__ = [
     "build_main_menu",
-    "generate_levels_keyboard",
-    "generate_terms_keyboard",
-    "generate_subjects_keyboard",
-    "generate_term_menu_keyboard_dynamic",
-    "generate_subject_sections_keyboard_dynamic",
-    "generate_section_filters_keyboard_dynamic",
-    "generate_years_keyboard",
-    "generate_lecturers_keyboard",
-    "generate_lecture_titles_keyboard",
-    "generate_lecturer_filter_keyboard",
-    "generate_year_category_menu_keyboard",
-    "generate_lecture_category_menu_keyboard",
-    "build_years_menu",
-    "build_lectures_menu",
-    "build_types_menu",
-    "build_exam_menu",
+    "build_children_keyboard",
 ]
-
-from .paginated import build_children_keyboard
-
-__all__ += ["build_children_keyboard"]
